+++ src/packages/piral-cli/src/commands.ts
@@ -84,9 +84,31 @@ export const allCommands: Array<ToolCommand<any>> = [
   {
     name: 'debug-pilet',
     alias: ['watch-pilet', 'debug', 'watch'],
-    description: '(currently not implemented)',
-    arguments: [],
-    run(args) {},
+    description: 'Starts the debugging process for a pilet using a Piral instance.',
+    arguments: ['[source]'],
+    flags(argv) {
+      return argv
+        .positional('source', {
+          type: 'string',
+          describe: 'Sets the source file containing the pilet root module.',
+          default: apps.debugPiletDefaults.entry,
+        })
+        .number('port')
+        .describe('port', 'Sets the port of the local development server.')
+        .default('port', apps.debugPiletDefaults.port)
+        .string('app')
+        .describe('app', 'Sets the name of the Piral instance.')
+        .string('base')
+        .default('base', process.cwd())
+        .describe('base', 'Sets the base directory. By default the current directory is used.');
+    },
+    run(args) {
+      return apps.debugPilet(args.base as string, {
+        entry: args.source as string,
+        port: args.port as number,
+        app: args.app as string,
+      });
+    },
   },
   {
     name: 'build-pilet',