+++ src/packages/piral-core/src/components/routes.test.tsx
@@ -20,6 +20,7 @@ jest.mock('../hooks');
         '/': StubComponent,
         '/custom': StubCustomPage,
       },
+      trackers: [],
     },
     components: {
       pages: {