+++ src/packages/piral/src/render.tsx
@@ -2,31 +2,33 @@ import * as React from 'react';
 import { render } from 'react-dom';
 import { RouteComponentProps } from 'react-router-dom';
 import { Provider } from 'urql';
-import { createInstance, PiralCoreApi, LocalizationMessages } from 'piral-core';
+import { createInstance, LocalizationMessages } from 'piral-core';
 import { createFetchApi, createGqlApi, setupGqlClient } from 'piral-ext';
 import { getGateway, getContainer, getAvailableModules } from './utils';
 import { getLayout } from './layout';
 import { getTranslations } from './translations';
 import { Loader, Dashboard, ErrorInfo } from './components';
-import { PiExtApi, PiletApi } from './api';
+import { PiExtApi, PiletApi, PiralAttachment } from './api';
 
 export interface PiralOptions {
   selector?: string | Element;
   gateway?: string;
   routes?: Record<string, React.ComponentType<RouteComponentProps>>;
+  trackers?: Array<React.ComponentType<RouteComponentProps>>;
   translations?: LocalizationMessages;
   components?: Record<string, React.ComponentType<any>>;
+  attach?: PiralAttachment;
 }
 
 export function renderInstance(options: PiralOptions = {}) {
-  const { routes = {}, selector = '#app', gateway, translations, components } = options;
+  const { routes = {}, trackers = [], selector = '#app', gateway, translations, components, attach } = options;
   const origin = getGateway(gateway);
   const client = setupGqlClient({
     url: origin,
   });
 
-  const Piral = createInstance({
-    availableModules: getAvailableModules(),
+  const Piral = createInstance<PiExtApi>({
+    availableModules: getAvailableModules(attach),
     requestModules() {
       return fetch(`${origin}/api/v1/pilet`)
         .then(res => res.json())
@@ -36,7 +38,7 @@ export function renderInstance(options: PiralOptions = {}) {
     Loader,
     Dashboard,
     ErrorInfo,
-    extendApi(api: PiralCoreApi<PiExtApi>): PiletApi {
+    extendApi(api): PiletApi {
       return {
         ...api,
         ...createFetchApi({
@@ -47,6 +49,7 @@ export function renderInstance(options: PiralOptions = {}) {
     },
     translations: getTranslations(translations),
     routes,
+    trackers,
   });
 
   const Layout = getLayout();